+++ LSLib/Granny/Model/CurveData/D9I3K8uC8u.cs
@@ -28,7 +28,7 @@ public class D9I3K8uC8u : AnimationCurveData
         var numKnots = NumKnots();
         var knots = new List<float>(numKnots);
         for (var i = 0; i < numKnots; i++)
-            knots.Add((float)KnotsControls[i] / scale);
+            knots.Add(KnotsControls[i] / scale);
 
         return knots;
     }
@@ -45,9 +45,9 @@ public class D9I3K8uC8u : AnimationCurveData
         for (var i = 0; i < numKnots; i++)
         {
             var mat = new Matrix3(
-                (float)KnotsControls[numKnots + i * 3 + 0] * ControlScales[0] + ControlOffsets[0], 0, 0,
-                0, (float)KnotsControls[numKnots + i * 3 + 1] * ControlScales[1] + ControlOffsets[1], 0,
-                0, 0, (float)KnotsControls[numKnots + i * 3 + 2] * ControlScales[2] + ControlOffsets[2]
+                KnotsControls[numKnots + i * 3 + 0] * ControlScales[0] + ControlOffsets[0], 0, 0,
+                0, KnotsControls[numKnots + i * 3 + 1] * ControlScales[1] + ControlOffsets[1], 0,
+                0, 0, KnotsControls[numKnots + i * 3 + 2] * ControlScales[2] + ControlOffsets[2]
             );
             knots.Add(mat);
         }
