+++ LSLib/LS/Story/Compiler/StoryEmitter.cs
@@ -57,7 +57,7 @@ public class StoryEmitter
 
             if (type.Value.TypeId == (uint)type.Value.IntrinsicTypeId)
             {
-                osiType.Alias = (byte)0;
+                osiType.Alias = 0;
                 osiType.IsBuiltin = true;
             }
             else
@@ -695,7 +695,7 @@ public class StoryEmitter
             {
                 if (!allowPartialPhysicalRow)
                 {
-                    adapter.LogicalIndices.Add((sbyte)-1);
+                    adapter.LogicalIndices.Add(-1);
                 }
             }
             else
@@ -720,14 +720,14 @@ public class StoryEmitter
                 var osiConst = EmitValue(param as IRConstant);
                 adapter.Constants.Physical.Add(osiConst);
                 adapter.Constants.Logical.Add(i, osiConst);
-                adapter.LogicalIndices.Add((sbyte)-1);
+                adapter.LogicalIndices.Add(-1);
             }
             else
             {
                 var variable = param as IRVariable;
                 if (rule.Variables[variable.Index].IsUnused())
                 {
-                    adapter.LogicalIndices.Add((sbyte)-1);
+                    adapter.LogicalIndices.Add(-1);
                 }
                 else
                 {
@@ -759,11 +759,11 @@ public class StoryEmitter
         else if (node is RelOpNode)
         {
             // (node as RelOpNode).AdapterRef.Resolve().LogicalIndices.Count
-            return EmitIdentityMappingAdapter(rule, (int)condition.TupleSize, true);
+            return EmitIdentityMappingAdapter(rule, condition.TupleSize, true);
         }
         else if (node is JoinNode)
         {
-            return EmitIdentityMappingAdapter(rule, (int)condition.TupleSize, true);
+            return EmitIdentityMappingAdapter(rule, condition.TupleSize, true);
         }
         else
         {
@@ -796,7 +796,7 @@ public class StoryEmitter
         Database db = null;
         if (left.DatabaseRef.IsValid && right.DatabaseRef.IsValid)
         {
-            db = EmitIntermediateDatabase(rule, (int)rightCondition.TupleSize, null);
+            db = EmitIntermediateDatabase(rule, rightCondition.TupleSize, null);
             if (db != null)
             {
                 database = new(Story, db);
@@ -925,7 +925,7 @@ public class StoryEmitter
         Database db = null;
         if (previousNode.DatabaseRef.IsValid)
         {
-            db = EmitIntermediateDatabase(rule, (int)condition.TupleSize, null);
+            db = EmitIntermediateDatabase(rule, condition.TupleSize, null);
             database = new(Story, db);
         }
         else
@@ -1037,7 +1037,7 @@ public class StoryEmitter
         Database db = null;
         if (previousNode.DatabaseRef.IsValid)
         {
-            db = EmitIntermediateDatabase(rule, (int)rule.Variables.Count, null);
+            db = EmitIntermediateDatabase(rule, rule.Variables.Count, null);
             if (db != null)
             {
                 database = new(Story, db);
